+++ net.adamec.ui.AppSwitcherBar/Win32/NativeMethods/DwmApi.cs
@@ -63,7 +63,18 @@ namespace net.adamec.ui.AppSwitcherBar.Win32.NativeMethods
         /// <param name="cbAttribute">The size, in bytes, of the attribute value being received via the pvAttribute parameter. The type of the retrieved value, and therefore its size in bytes, depends on the value of the dwAttribute parameter.</param>
         /// <returns>If the function succeeds, it returns S_OK. Otherwise, it returns an HRESULT error code.</returns>
         [DllImport(DLL_NAME)]
-        public static extern int DwmGetWindowAttribute(IntPtr hwnd, uint dwAttribute, out bool pvAttribute, int cbAttribute);
+        public static extern int DwmGetWindowAttribute(IntPtr hwnd, uint dwAttribute, ref bool pvAttribute, int cbAttribute);
+
+        /// <summary>
+        /// Retrieves the current value of a specified Desktop Window Manager (DWM) attribute applied to a window.
+        /// </summary>
+        /// <param name="hwnd">The handle to the window from which the attribute value is to be retrieved.</param>
+        /// <param name="dwAttribute">A flag describing which value to retrieve, specified as a value of the DWMWINDOWATTRIBUTE enumeration. This parameter specifies which attribute to retrieve, and the pvAttribute parameter points to an object into which the attribute value is retrieved.</param>
+        /// <param name="pvAttribute">A pointer to a value which, when this function returns successfully, receives the current value of the attribute. The type of the retrieved value depends on the value of the dwAttribute parameter. The DWMWINDOWATTRIBUTE enumeration topic indicates, in the row for each flag, what type of value you should pass a pointer to in the pvAttribute parameter.</param>
+        /// <param name="cbAttribute">The size, in bytes, of the attribute value being received via the pvAttribute parameter. The type of the retrieved value, and therefore its size in bytes, depends on the value of the dwAttribute parameter.</param>
+        /// <returns>If the function succeeds, it returns S_OK. Otherwise, it returns an HRESULT error code.</returns>
+        [DllImport(DLL_NAME)]
+        public static extern int DwmGetWindowAttribute(IntPtr hwnd, uint dwAttribute, ref int pvAttribute, int cbAttribute);
 
 
     }